MDEV-35299: Server build failure on MacOS Monterey
Build failure was caused by using the sprintf C API function
that is marked as deprecated in MacOS Monterey. Since building
of server is performaed with the compiler flag -Werror
it results in build failure.

To fix the issue, every use of the C API function sprintf
has been replaced with snprintf throughout the source code.
Source code of the storage engine CONNECT wasn't modified.