MDEV-32294 2nd execution problem with inconsistent outer context paths
WORK IN PROGRESS

context {select_lex, outer_context} elements made private, accessor
function will now skip merged selects.

When calculating the Item_subselect::used_tables_cache the current code
relies on upper_refs, this is populated by Item_subselect::mark_as_dependent()
which gets called during fix_fields on a Field/Reference that is an
outer reference.
This list contains an item contained within the Item_subselect somewhere
and a select_lex where that item is resolved.
For each item in this list, Item_subselect::update_used_tables() searches
outwards until it encounters 'new_parent'.  If it does encounter this
select_lex, then the tables used in this item are records as being used
in this Item_subselect.  If it doesn't encounter this select_lex, then
it sets OUTER_BIT_TABLE_REF indicating that the somewhere within lies
and item resolved above our new parent.

This new code code turns this upside down by directly referencing items
resolved in a SELECT_LEX, in that SELECT_LEX.  Later, when recalculating
Item_subselect::used_tables_cache, rather than looking though all items
below that might be resolved within the SELECT_LEX of interest
(new_parent), we can iterate through the items resolved in 'new_parent'
and decide if they are relevant to our Item_subquery.  We do this by
starting in the select_lex where the item is defined, walking outwards
on outer_select() and seeing if the master_unit() is the same as
Item_subselect::unit.  If it is, this item 'belongs' to our
Item_subselect.

Changes:

We add SELECT_LEX::resolved_here    List<Item>
Item because we can gather Item_field or Item_direct_view_ref
We add to this list on statement memory only during prepare or first execution.

When looking at items in 'new_parent' we can determine where this item is
used in this Item_subselect by walking outwards from where the item is
defined up to 'new_parent' looking to see if the enclosing unit is the
unit attached to out Item_subselect.
If we have a match then we know that we add the bit
field_item->field->table->map into our bitmap (used_tables_cache).

Test to see if this Item belongs to our unit implemented in
Item_belongs_to()

When we merge 2 queries together (either a derived table merge or a
subquery to semijoin merge), we need to manage the resolved_here list
for both SELECT_LEXs.
The SELECT_LEX that is uppermost must have it's resolved_here list
pruned of items that are belong to the SELECT_LEX to be merged.  These
pruned items will no longer be outer references.